requirements.txt
python-telegram-bot==12.5.1
requests==2.23.0
